Renamed classes and made utils for email/password regex
