--- src/main/java/dev/devature/penguin_api/interfaces/AuthenticationStrategy.java
@@ -1,8 +1,0 @@
-package dev.devature.penguin_api.interfaces;
-
-import dev.devature.penguin_api.dto.AuthRequest;
-
-public interface AuthenticationStrategy {
-    boolean checkValidity(AuthRequest authRequest);
-    boolean authenticate(AuthRequest authRequest);
-}
