--- src/main/resources/application-test.properties
@@ -1,9 +1,0 @@
-
-# H2 database connection properties
-spring.datasource.url=jdbc:h2:mem:testdb;MODE=PostgreSQL
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.h2.console.enabled=true
-
